+++ tests/linter/test_format_checks.py
@@ -1,6 +1,7 @@
 import pytest
 
 from .common import simple_ok_check, simple_nok_check
+from gdtoolkit.linter import lint_code, DEFAULT_CONFIG
 
 
 def test_max_file_lines_ok():
@@ -18,6 +19,12 @@ def test_max_file_lines_nok():
 """
 #xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
 """,
+"""
+func foo():
+	if true:
+		if true:
+			var x = 1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1
+""",
 ])
 def test_max_line_length_ok(code):
     simple_ok_check(code)
@@ -30,6 +37,24 @@ def test_max_line_length_ok(code):
 ])
 def test_max_line_length_nok(code):
     simple_nok_check(code, 'max-line-length')
+
+
+@pytest.mark.parametrize('code', [
+"""
+func foo():
+	if true:
+		if true:
+			var x = 1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1+1
+""",
+])
+def test_max_line_length_w_tabs_as_8_nok(code):
+    config = DEFAULT_CONFIG.copy()
+    config.update({"tab-characters": 8})
+    outcome = lint_code(code, config)
+    assert len(outcome) == 1
+    assert outcome[0].name == 'max-line-length'
+    assert outcome[0].line == 5
+
 
 
 @pytest.mark.parametrize('code', [