+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Added a possibility to disable linter checks in range from e.g. `# gdlint: disable=function-name` to `# gdlint: enable=function-name`
  - Added support for `puppet var`
  - Added string formatting according to `GDScript style guide` (`'x'` -> `"x"`, `'"'` -> `'"'`. `"'"` -> `"'"`, `'"\''` -> `"\"'"`)
+ - Added linter option `tab-characters` that allows treating tabs as multiple characters while calculating line length
 
 ### Fixed
  - Fixed comment persistence check